refactor network call on follower list screen to manage memory by capturing weak self
