+++ GHFollowers/GHFollowers/Controllers/FollowerListVC.swift
@@ -40,7 +40,7 @@ class FollowerListVC: UIViewController {
   }
   
   func configureCollectionView() {
-    collectionView = UICollectionView(frame: view.bounds, collectionViewLayout: createThreeColumnFlowLayout())
+    collectionView = UICollectionView(frame: view.bounds, collectionViewLayout: UIHelper.createThreeColumnFlowLayout(in: view))
     view.addSubview(collectionView)
     collectionView.backgroundColor = .systemBackground
     collectionView.register(FollowerCell.self, forCellWithReuseIdentifier: FollowerCell.reuseIdentifier)
@@ -61,22 +61,9 @@ class FollowerListVC: UIViewController {
     dataSource.apply(snapshot, animatingDifferences: true)
   }
   
-  func createThreeColumnFlowLayout() -> UICollectionViewFlowLayout {
-    let width = view.bounds.width
-    let padding: CGFloat = 12
-    let minimumItemSpacing: CGFloat = 10
-    let availableWidth = width - (padding * 2) - (minimumItemSpacing * 2)
-    let itemWidth = availableWidth / 3
-    
-    let flowLayout = UICollectionViewFlowLayout()
-    flowLayout.sectionInset = UIEdgeInsets(top: padding, left: padding, bottom: padding, right: padding)
-    flowLayout.itemSize = CGSize(width: itemWidth, height: itemWidth + 40)
-    
-    return flowLayout
-  }
-  
   func getFollowers() {
-    NetworkManager.shared.getFollowers(for: username, page: 1, perPageFollowers: 100) { result in
+    NetworkManager.shared.getFollowers(for: username, page: 1, perPageFollowers: 100) { [weak self] result in
+      guard let self = self else { return }
       switch result {
       case .failure(let error):
         self.presentGFAlertOnMainThread(title: "ERROR!!!", message: error.rawValue, buttonTitle: "OK")
