+++ GHFollowers/GHFollowers/Custom Views/Cells/FollowerCell.swift
@@ -35,14 +35,14 @@ class FollowerCell: UICollectionViewCell, ReuseIdentifying {
     addSubview(usernameLabel)
     
     NSLayoutConstraint.activate([
-      avatarImageView.topAnchor.constraint(equalTo: contentView.topAnchor, constant: padding),
-      avatarImageView.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: padding),
-      avatarImageView.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -padding),
+      avatarImageView.topAnchor.constraint(equalTo: topAnchor, constant: padding),
+      avatarImageView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: padding),
+      avatarImageView.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -padding),
       avatarImageView.heightAnchor.constraint(equalTo: avatarImageView.widthAnchor),
       
       usernameLabel.topAnchor.constraint(equalTo: avatarImageView.bottomAnchor, constant: 12),
-      usernameLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: padding),
-      usernameLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -padding),
+      usernameLabel.leadingAnchor.constraint(equalTo: leadingAnchor, constant: padding),
+      usernameLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -padding),
       usernameLabel.heightAnchor.constraint(equalToConstant: 20)
     ])
   }
